// frontend/src/generated/models/decisionRunResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { DecisionRunResponseApprovedBy } from './decisionRunResponseApprovedBy';
import type { DecisionContextResponse } from './decisionContextResponse';
import type { DecisionRunKind } from './decisionRunKind';
import type { DecisionRunResponseParameters } from './decisionRunResponseParameters';
import type { DecisionRunResponseRejectedBy } from './decisionRunResponseRejectedBy';
import type { DecisionRunResponseRejectionReason } from './decisionRunResponseRejectionReason';
import type { DecisionRunResponseResult } from './decisionRunResponseResult';
import type { DecisionRunResponseValidatedBy } from './decisionRunResponseValidatedBy';

export interface DecisionRunResponse {
  approved_by?: DecisionRunResponseApprovedBy;
  context: DecisionContextResponse;
  created_at: string;
  created_by: string;
  kind: DecisionRunKind;
  model_version: string;
  parameters: DecisionRunResponseParameters;
  projection_version: number;
  references: string[];
  rejected_by?: DecisionRunResponseRejectedBy;
  rejection_reason?: DecisionRunResponseRejectionReason;
  result: DecisionRunResponseResult;
  run_id: string;
  source_snapshot_ids: string[];
  status: string;
  validated_by?: DecisionRunResponseValidatedBy;
}